# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error: {}
    type: object
  models.FacturaDatosBesser:
    properties:
      direccion:
        type: string
      docid:
        type: string
      email:
        type: string
      telefono:
        type: string
    type: object
  models.FacturaDetResponse:
    properties:
      cantidad:
        type: number
      concepto:
        type: string
      precio_unitario:
        $ref: '#/definitions/models.Moneda'
      suscripcion_detalle:
        additionalProperties: {}
        type: object
      tax_status:
        type: string
      total:
        $ref: '#/definitions/models.Moneda'
    type: object
  models.FacturaList:
    properties:
      created_at:
        type: string
      estatus:
        type: string
      factura_id:
        type: string
      monto_pagado:
        $ref: '#/definitions/models.Moneda'
      monto_total:
        $ref: '#/definitions/models.Moneda'
      nreferencia:
        type: string
    type: object
  models.FacturaResponse:
    properties:
      base_imponible:
        $ref: '#/definitions/models.Moneda'
      created_at:
        type: string
      datos_besser:
        $ref: '#/definitions/models.FacturaDatosBesser'
      desc_monto_dolar:
        $ref: '#/definitions/models.Moneda'
      desc_porcentaje:
        type: number
      direccion:
        type: string
      docid:
        type: string
      estatus:
        type: string
      factura_detalle:
        items:
          $ref: '#/definitions/models.FacturaDetResponse'
        type: array
      factura_id:
        type: string
      igtf_base_imponible:
        $ref: '#/definitions/models.Moneda'
      igtf_monto:
        $ref: '#/definitions/models.Moneda'
      igtf_porcentaje:
        type: number
      iva_monto:
        $ref: '#/definitions/models.Moneda'
      iva_porcentaje:
        type: number
      nreferencia:
        type: string
      pagos:
        items:
          $ref: '#/definitions/models.PaymentList'
        type: array
      razon_social:
        type: string
      retenciones:
        items:
          $ref: '#/definitions/models.RetencionList'
        type: array
      subtotal:
        $ref: '#/definitions/models.Moneda'
      telefono:
        type: string
      total:
        $ref: '#/definitions/models.Moneda'
    type: object
...
  models.FormaPagoList:
    properties:
      banco:
        type: string
      detalle: {}
      id:
        type: integer
      metodo_pago:
        type: string
      moneda:
        type: string
      nombre:
        type: string
    type: object
...
  models.Moneda:
    properties:
      bolivar:
        type: number
      dolar:
        type: number
    type: object
  models.PaginatorData:
    properties:
      current_page:
        type: integer
      next_page:
        type: integer
      prev_page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.PaymentList:
    properties:
      created_at:
        type: string
      estatus:
        type: string
      fecha:
        type: string
      metodo_pago:
        $ref: '#/definitions/models.FormaPagoList'
      monto_total:
        $ref: '#/definitions/models.Moneda'
      ncontrol:
        type: string
      payment_id:
        type: string
      referencia:
        type: string
    type: object
...
  models.RetencionList:
    properties:
      created_at:
        type: string
      estatus:
        type: string
      fecha_retencion:
        type: string
      monto_retenido:
        $ref: '#/definitions/models.Moneda'
      ncontrol:
        type: string
      num_comprobante:
        type: string
      numero_factura:
        type: string
      retencion_id:
        type: string
      tipo_retencion:
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      notice:
        type: string
      record: {}
    type: object
  models.SuccessResponseWithMeta:
    properties:
      _meta:
        $ref: '#/definitions/models.PaginatorData'
      notice:
        type: string
      record: {}
    type: object
...
host: 127.0.0.1:7003
info:
  contact:
    email: juan9acosta@gmail.com
    name: Juan Acosta
    url: https://www.linkedin.com/in/juan-m-acosta-f-54219758/
  description: service in Go using Gin framework
  title: MiCuenta Service API
  version: "1.0"
paths:
...
  /cron/clean-old-sessions:
    get:
      consumes:
      - application/json
      description: cleaning old session of clientes from postgresDB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_users
      tags:
      - Crons
  /cron/create-clients-passwd:
    get:
      consumes:
      - application/json
      description: search for all users with password null and set the password the
        same as the docid
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task create_client_passwd
      tags:
      - Crons
  /cron/sinc-factura-fiscal:
    get:
      consumes:
      - application/json
      description: busca registros nuevos en la bd de mysql y sincroniza la data a
        postgres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_factura_fiscal
      tags:
      - Crons
  /cron/sinc-prefactura-anulada:
    get:
      consumes:
      - application/json
      description: busca registros nuevos en la bd de mysql y sincroniza la data a
        postgres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_prefactura_anuladas
      tags:
      - Crons
  /cron/sinc-prefactura-pagadas:
    get:
      consumes:
      - application/json
      description: busca registros nuevos en la bd de mysql y sincroniza la data a
        postgres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_prefactura_pagadas
      tags:
      - Crons
  /cron/sinc-recibov-anulado:
    get:
      consumes:
      - application/json
      description: busca registros nuevos en la bd de mysql y sincroniza la data a
        postgres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_recibov_anulado
      tags:
      - Crons
  /cron/sinc-recibov-procesado:
    get:
      consumes:
      - application/json
      description: busca registros nuevos en la bd de mysql y sincroniza la data a
        postgres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_recibov_procesado
      tags:
      - Crons
  /cron/sinc-retencion:
    get:
      consumes:
      - application/json
      description: busca registros nuevos en la bd de mysql y sincroniza la data a
        postgres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_retenciones
      tags:
      - Crons
  /cron/sinc-tasa-cambio:
    get:
      consumes:
      - application/json
      description: busca registros nuevos en la bd de mysql y sincroniza la data a
        postgres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BasicAuth: []
      summary: Run the task sinc_tasa_cambio
      tags:
      - Crons
  /factura/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of facturas with pagination
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of recibos de pago per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponseWithMeta'
            - properties:
                record:
                  items:
                    $ref: '#/definitions/models.FacturaList'
                  type: array
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listado de facturas
      tags:
      - Factura
  /factura/show:
    get:
      consumes:
      - application/json
      description: devuelve toda la data relacionada a una factura
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: retencionId (UUID), created_at(timestamptz)
        in: query
        name: FacturaReqId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  $ref: '#/definitions/models.FacturaResponse'
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: detalle de una factura
      tags:
      - Factura
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"